rename issue in previous commit, overwriting with new temp name
